tas256x: Read ChipID

+++ tas256X_win/Device.h
@@ -16,9 +16,9 @@ typedef struct _SPB_CONTEXT
 typedef struct _DEVICE_CONTEXT
 {
     SPB_CONTEXT             SpbContextA;
-    INT                     SpbContextA_ID;
+    UINT8                   SpbContextA_ID;
     SPB_CONTEXT             SpbContextB;
-    INT                     SpbContextB_ID;
+    UINT8                   SpbContextB_ID;
     BOOLEAN                 TwoSpeakers;
     PCALLBACK_OBJECT        CSAudioAPICallback;
     PVOID                   CSAudioAPICallbackObj;
